fix: review comments and some refactorings
