+++ test/library/validation-test.ts
@@ -1,71 +1,62 @@
 // /test/library/validation-test.ts
 // Tests the validation functions
 
-import {
-	afterEach,
-	beforeEach,
-	describe,
-	expect,
-	it,
-	jest,
-} from '@jest/globals'
+import { beforeEach, describe, expect, it, jest } from '@jest/globals'
 import express from 'express'
 import supertest from 'supertest'
 import { ipKeyGenerator, MemoryStore } from '../../source/index.js'
-import type { Logger } from '../../source/logger'
-import type { Store } from '../../source/types'
+import type { Logger, Store } from '../../source/types'
 import { getValidations, type Validations } from '../../source/validations.js'
 
 describe('validations tests', () => {
 	let validations: Validations
+	let logger: Logger
 
 	beforeEach(() => {
-		validations = getValidations(true)
+		logger = {
+			warn: jest.fn(),
+			error: jest.fn(),
+		}
 
-		jest.spyOn(console, 'error').mockImplementation(() => {})
-		jest.spyOn(console, 'warn').mockImplementation(() => {})
-	})
-
-	afterEach(() => {
-		jest.restoreAllMocks()
+		validations = getValidations(true, logger)
 	})
 
 	describe('ip', () => {
 		it('should allow a valid IPv4', () => {
 			validations.ip('1.2.3.4')
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should allow a valid IPv6', () => {
 			validations.ip('1200:0000:AB00:1234:0000:2552:7777:1313')
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should log an error for an invalid IP', () => {
 			validations.ip('badip')
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should log an error for an undefined IP', () => {
 			validations.ip(undefined)
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should log an error for an IPv4 with a port', () => {
 			validations.ip('1.2.3.4:1234')
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should log an error for an IPv6 with a port', () => {
 			validations.ip('[1200:0000:AB00:1234:0000:2552:7777:1313]:1234')
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 	})
 
 	describe('trustProxy', () => {
 		it('should log an error on "trust proxy" = true', () => {
 			validations.trustProxy({ app: { get: () => true } } as any)
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should not log an error on "trust proxy" != true', () => {
@@ -73,7 +64,7 @@ describe('validations tests', () => {
 			validations.trustProxy({ app: { get: () => '1.2.3.4' } } as any)
 			validations.trustProxy({ app: { get: () => /1.2.3.4/ } } as any)
 			validations.trustProxy({ app: { get: () => ['1.2.3.4'] } } as any)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
@@ -91,13 +82,13 @@ describe('validations tests', () => {
 				app: { get: () => false },
 				headers: {},
 			} as any)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 
 			validations.xForwardedForHeader({
 				app: { get: () => false },
 				headers: { 'x-forwarded-for': '1.2.3.4' },
 			} as any)
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 	})
 
@@ -108,14 +99,14 @@ describe('validations tests', () => {
 				ip: '1.2.3.4',
 				socket: { remoteAddress: '1.2.3.4' },
 			} as any)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 
 			validations.forwardedHeader({
 				headers: { forwarded: '5.6.7.8' },
 				ip: '1.2.3.4',
 				socket: { remoteAddress: '1.2.3.4' },
 			} as any)
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_FORWARDED_HEADER' }),
 			)
 		})
@@ -126,36 +117,36 @@ describe('validations tests', () => {
 				ip: '1.2.3.100',
 				socket: { remoteAddress: '1.2.3.4' },
 			} as any)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 
 			validations.forwardedHeader({
 				headers: { forwarded: '5.6.7.8' },
 				ip: '1.2.3.100',
 				socket: { remoteAddress: '1.2.3.4' },
 			} as any)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
 	describe('positiveHits', () => {
 		it('should log an error if hits is non-numeric', () => {
 			validations.positiveHits(true)
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should log an error if hits is less than 1', () => {
 			validations.positiveHits(0)
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should log an error if hits is not an integer', () => {
 			validations.positiveHits(1.5)
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should not log an error if hits is a positive integer', () => {
 			validations.positiveHits(1)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
@@ -164,9 +155,9 @@ describe('validations tests', () => {
 			const store = { localKeys: true }
 
 			validations.unsharedStore(store as Store)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 			validations.unsharedStore(store as Store)
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_STORE_REUSE',
 				}),
@@ -177,9 +168,9 @@ describe('validations tests', () => {
 			const store = { localKeys: false }
 
 			validations.unsharedStore(store as Store)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 			validations.unsharedStore(store as Store)
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_STORE_REUSE',
 					message: expect.stringContaining('unique prefix'),
@@ -193,7 +184,7 @@ describe('validations tests', () => {
 
 			validations.unsharedStore(store1 as Store)
 			validations.unsharedStore(store2 as Store)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
@@ -208,9 +199,9 @@ describe('validations tests', () => {
 			const key = '1.2.3.4'
 
 			validations.singleCount(request as any, store as Store, key)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 			validations.singleCount(request as any, store as Store, key)
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_DOUBLE_COUNT',
 				}),
@@ -223,9 +214,9 @@ describe('validations tests', () => {
 			const key = '1.2.3.4'
 
 			validations.singleCount(request as any, store as Store, key)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 			validations.singleCount(request as any, store as Store, key)
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_DOUBLE_COUNT',
 				}),
@@ -240,7 +231,7 @@ describe('validations tests', () => {
 
 			validations.singleCount(request as any, store1 as Store, key)
 			validations.singleCount(request as any, store2 as Store, key)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should log an error if a request is double-counted with separate instances of an external store', () => {
@@ -251,7 +242,7 @@ describe('validations tests', () => {
 
 			validations.singleCount(request as any, store1 as Store, key)
 			validations.singleCount(request as any, store2 as Store, key)
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_DOUBLE_COUNT',
 				}),
@@ -265,9 +256,9 @@ describe('validations tests', () => {
 			const key = '1.2.3.4'
 
 			validations.singleCount(request1 as any, store as Store, key)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 			validations.singleCount(request2 as any, store as Store, key)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should not log an error if a request is double-counted with separate instances of an external store with different prefixes', () => {
@@ -280,82 +271,82 @@ describe('validations tests', () => {
 
 			validations.singleCount(request as any, store1 as Store, key)
 			validations.singleCount(request as any, store2 as Store, key)
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
 	describe('limit', () => {
 		it('should log a warning if max is set to 0', () => {
 			validations.limit(0)
-			expect(console.warn).toHaveBeenCalled()
+			expect(logger.warn).toHaveBeenCalled()
 		})
 
 		it('should not log a warning if max is set to a non zero number', () => {
 			validations.limit(3)
-			expect(console.warn).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
 		})
 	})
 
 	describe('onLimitReached', () => {
 		it('should log a warning if onLimitReached is set', () => {
 			validations.onLimitReached(() => {})
-			expect(console.warn).toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should not log a warning if onLimitReached is unset', () => {
 			validations.onLimitReached(undefined)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
 	describe('draft_polli_ratelimit_headers', () => {
 		it('should log a warning if draft_polli_ratelimit_headers is set', () => {
 			validations.draftPolliHeaders(true)
-			expect(console.warn).toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should not log a warning if draft_polli_ratelimit_headers is unset or false', () => {
 			validations.draftPolliHeaders(false)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 
 			validations.draftPolliHeaders(undefined)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
 	describe('headersDraftVersion', () => {
 		it('should log an error if standardHeaders is an unsupported version', () => {
 			validations.headersDraftVersion('draft-2')
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should not log an error a valid version is passed as standardHeaders', () => {
 			validations.headersDraftVersion('draft-8')
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
 	describe('headersResetTime', () => {
 		it('should log an error if resetTime is omitted', () => {
 			validations.headersResetTime(undefined)
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 
 		it('should not log an error if resetTime is set', () => {
 			validations.headersResetTime(new Date())
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
 	describe('knownOptions', () => {
 		it('should log an error if an unknown option is passed in', () => {
 			validations.knownOptions({ windowMS: 100 } as any)
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_UNKNOWN_OPTION' }),
 			)
 		})
@@ -363,56 +354,68 @@ describe('validations tests', () => {
 
 	describe('validationsConfig', () => {
 		it('should log an error if an unknown validation is disabled', () => {
-			validations = getValidations({ invalid: false } as any)
+			validations = getValidations({ invalid: false } as any, logger)
 
 			validations.validationsConfig()
-			expect(console.error).toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalled()
 		})
 		it('should log an error if an unknown validation is enabled', () => {
-			validations = getValidations({ invalid: false } as any)
+			validations = getValidations({ invalid: false } as any, logger)
 
 			validations.validationsConfig()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_UNKNOWN_VALIDATION' }),
 			)
 		})
 		it('should not log an error if only valid keys are set', () => {
-			validations = getValidations({
-				ip: false,
-				positiveHits: true,
-				default: false,
-			})
+			validations = getValidations(
+				{
+					ip: false,
+					positiveHits: true,
+					default: false,
+				},
+				logger,
+			)
 
 			validations.validationsConfig()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 		it('should not run if disabled by config', () => {
-			validations = getValidations({
-				invalid: false,
-				validationsConfig: false,
-			} as any)
+			validations = getValidations(
+				{
+					invalid: false,
+					validationsConfig: false,
+				} as any,
+				logger,
+			)
 
 			validations.validationsConfig()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 		it('should not run if disabled by default', () => {
-			validations = getValidations({
-				invalid: true,
-				default: false,
-			} as any)
+			validations = getValidations(
+				{
+					invalid: true,
+					default: false,
+				} as any,
+				logger,
+			)
 
 			validations.validationsConfig()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 		it('should run if enabled by config with default: false', () => {
-			validations = getValidations({
-				invalid: false,
-				validationsConfig: true,
-				default: false,
-			} as any)
+			validations = getValidations(
+				{
+					invalid: false,
+					validationsConfig: true,
+					default: false,
+				} as any,
+				logger,
+			)
 
 			validations.validationsConfig()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_UNKNOWN_VALIDATION' }),
 			)
 		})
@@ -420,17 +423,17 @@ describe('validations tests', () => {
 
 	describe('disable', () => {
 		it('should initialize disabled when passed false', () => {
-			const disabledValidator = getValidations(false)
+			const disabledValidator = getValidations(false, logger)
 
 			disabledValidator.ip('badip')
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should do nothing after disable() is called', () => {
 			validations.disable()
 
 			validations.ip('badip')
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
@@ -445,12 +448,12 @@ describe('validations tests', () => {
 			})
 
 			await supertest(app).get('/').expect('hello')
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_CREATED_IN_REQUEST_HANDLER',
 				}),
 			)
-			expect(console.warn).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
 		})
 
 		it('should log a different error when used with an external store', async () => {
@@ -463,20 +466,20 @@ describe('validations tests', () => {
 			})
 
 			await supertest(app).get('/').expect('hello')
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_CREATED_IN_REQUEST_HANDLER',
 				}),
 			)
-			expect(console.warn).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
 		})
 
 		it('should not log an error if called elsewhere', async () => {
 			const store = new MemoryStore()
 
 			validations.creationStack(store)
-			expect(console.error).not.toHaveBeenCalled()
-			expect(console.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
 		})
 	})
 
@@ -486,44 +489,44 @@ describe('validations tests', () => {
 				validations.ipv6Subnet(i)
 			}
 
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should allow false', () => {
 			validations.ipv6Subnet(false)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should error on numbers below 32', () => {
 			validations.ipv6Subnet(31)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_IPV6_SUBNET' }),
 			)
 		})
 
 		it('should error on numbers above 64', () => {
 			validations.ipv6Subnet(65)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_IPV6_SUBNET' }),
 			)
 		})
 
 		it('should error on non-integer numbers', () => {
 			validations.ipv6Subnet(48.5)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_IPV6_SUBNET' }),
 			)
 		})
 
 		it('should error on undefined (return value from configured function)', () => {
 			validations.ipv6Subnet(undefined)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_IPV6_SUBNET' }),
 			)
 		})
@@ -536,8 +539,8 @@ describe('validations tests', () => {
 			validations.ipv6SubnetOrKeyGenerator({
 				keyGenerator: (request, response) => 'global',
 			})
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should warn on both', () => {
@@ -545,8 +548,8 @@ describe('validations tests', () => {
 				ipv6Subnet: 64,
 				keyGenerator: (request, response) => 'global',
 			})
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_IPV6SUBNET_OR_KEYGENERATOR',
 				}),
@@ -558,8 +561,8 @@ describe('validations tests', () => {
 				ipv6Subnet: false,
 				keyGenerator: (request, response) => 'global',
 			})
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({
 					code: 'ERR_ERL_IPV6SUBNET_OR_KEYGENERATOR',
 				}),
@@ -570,8 +573,8 @@ describe('validations tests', () => {
 	describe('keyGeneratorIpFallback', () => {
 		it('should skip on undefined keyGenerator', () => {
 			validations.keyGeneratorIpFallback(undefined)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should not warn on a keyGenerator that does not use req.ip or request.ip', () => {
@@ -579,8 +582,8 @@ describe('validations tests', () => {
 				(request: any, response: any): string =>
 					request.params.apikey as string,
 			)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 
 		it('should warn on a keyGenerator that uses req.ip', () => {
@@ -588,8 +591,8 @@ describe('validations tests', () => {
 				(request: any, response: any): string =>
 					(request.params.apikey || request.ip) as string,
 			)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_KEY_GEN_IPV6' }),
 			)
 		})
@@ -599,8 +602,8 @@ describe('validations tests', () => {
 				(request: any, response: any) =>
 					(request.params.apikey || request.ip) as string,
 			)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_KEY_GEN_IPV6' }),
 			)
 		})
@@ -610,46 +613,21 @@ describe('validations tests', () => {
 				(request: any, response: any): string =>
 					(request.params.apikey as string) || ipKeyGenerator(request.ip),
 			)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 		})
 	})
 
 	describe('windowMs', () => {
 		it('should warn on large values, but not in-range values', () => {
 			validations.windowMs(5 * 60 * 1000)
-			expect(console.warn).not.toHaveBeenCalled()
-			expect(console.error).not.toHaveBeenCalled()
+			expect(logger.warn).not.toHaveBeenCalled()
+			expect(logger.error).not.toHaveBeenCalled()
 
 			validations.windowMs(30 * 24 * 60 * 60 * 1000)
-			expect(console.error).toHaveBeenCalledWith(
+			expect(logger.error).toHaveBeenCalledWith(
 				expect.objectContaining({ code: 'ERR_ERL_WINDOW_MS' }),
 			)
-		})
-	})
-
-	describe('with custom logger implementation', () => {
-		let logger: Logger
-
-		beforeEach(() => {
-			logger = {
-				warn: jest.fn(),
-				error: jest.fn(),
-			}
-
-			validations = getValidations(true, logger)
-		})
-
-		it('calls logger.error on error', () => {
-			validations.ip('badip')
-			expect(logger.error).toHaveBeenCalledWith(expect.any(Error))
-			expect(console.error).not.toHaveBeenCalled()
-		})
-
-		it('calls logger.warn on warning', () => {
-			validations.limit(0)
-			expect(logger.warn).toHaveBeenCalledWith(expect.any(Error))
-			expect(console.warn).not.toHaveBeenCalled()
 		})
 	})
 })
